prewarm_lfc review pass
