+++ compute_tools/src/http/routes/prewarm_lfc.rs
@@ -1,49 +1,38 @@
-use crate::compute::{ComputeNode, EndpointStoragePair, PrewarmStatus};
-use axum::response::{IntoResponse, Response};
+use crate::http::JsonResponse;
+use axum::response::Response;
 use axum::{Json, http::StatusCode};
 use compute_api::responses::PrewarmOffloadState;
-use std::sync::Arc;
-use tracing::error;
+type Compute = axum::extract::State<std::sync::Arc<crate::compute::ComputeNode>>;
 
-type State = axum::extract::State<Arc<ComputeNode>>;
+pub(in crate::http) async fn state(compute: Compute) -> Json<crate::compute_prewarm::PrewarmState> {
+    Json(compute.prewarm_state().await)
+}
 
-impl axum::extract::FromRequestParts<Arc<ComputeNode>> for EndpointStoragePair {
-    type Rejection = Response;
-    async fn from_request_parts(
-        _: &mut http::request::Parts,
-        state: &Arc<ComputeNode>,
-    ) -> Result<Self, Self::Rejection> {
-        let Some(ref pspec) = state.state.lock().unwrap().pspec else {
-            error!("pspec missing");
-            return Err(StatusCode::BAD_REQUEST.into_response());
-        };
-        pspec.try_into().map_err(|e| {
-            error!("{e}");
-            StatusCode::BAD_REQUEST.into_response()
-        })
+// Following functions are marked async for axum, as it's more convenient than wrapping these
+// in async lambdas at call site
+
+pub(in crate::http) async fn offload_state(compute: Compute) -> Json<PrewarmOffloadState> {
+    Json(compute.prewarm_offload_state())
+}
+
+pub(in crate::http) async fn prewarm(compute: Compute) -> Response {
+    if compute.prewarm() {
+        JsonResponse::success(StatusCode::OK, "")
+    } else {
+        JsonResponse::error(
+            StatusCode::TOO_MANY_REQUESTS,
+            "Multiple requests for prewarm are not allowed",
+        )
     }
 }
 
-pub(in crate::http) async fn status(state: State) -> Json<PrewarmStatus> {
-    Json(state.prewarm_status().await)
-}
-
-pub(in crate::http) async fn offload_status(state: State) -> Json<PrewarmOffloadState> {
-    Json(state.prewarm_offload_status().await)
-}
-
-pub(in crate::http) async fn prewarm(state: State, pair: EndpointStoragePair) -> StatusCode {
-    if state.prewarm(pair).await {
-        StatusCode::OK
+pub(in crate::http) async fn offload(compute: Compute) -> Response {
+    if compute.prewarm_offload() {
+        JsonResponse::success(StatusCode::OK, "")
     } else {
-        StatusCode::TOO_MANY_REQUESTS
-    }
-}
-
-pub(in crate::http) async fn offload(state: State, pair: EndpointStoragePair) -> StatusCode {
-    if state.prewarm_offload(pair).await {
-        StatusCode::OK
-    } else {
-        StatusCode::TOO_MANY_REQUESTS
+        JsonResponse::error(
+            StatusCode::TOO_MANY_REQUESTS,
+            "Multiple requests for prewarm offload are not allowed",
+        )
     }
 }
