+++ control_plane/src/bin/neon_local.rs
@@ -1491,7 +1491,7 @@ async fn handle_endpoint(subcmd: &EndpointCmd, env: &local_env::LocalEnv) -> Res
             let exp = (std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH)?
                 + Duration::from_secs(86400))
             .as_secs();
-            let claims = endpoint_storage::claims::EndpointClaims {
+            let claims = endpoint_storage::claims::EndpointStorageClaims {
                 tenant_id: endpoint.tenant_id,
                 timeline_id: endpoint.timeline_id,
                 endpoint_id: endpoint_id.to_string(),
